Add Screen Brightness Helpers
The script ux581_brightness.sh reads the currently requested brightness
and the maximum brightness and generates a percent representation. There
is a min/max percent brightness for both the OLED and Screenpad displays
and the percentage is scaled accordingly. The Screenpad brightness is
then scaled to 0-255 (0x00-0xFF) for the /proc/acpi/call request that
Plippo documented in
s-light#1

The path and service files have been included for SystemD to set up
inotifywatch on the brightness file and an INSTALL text file has been
included to show where to place the files and what to run to enable them
on startup.